add Makefile & modifing upto strlcpy

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: taehykim <taehykim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/23 16:35:34 by taehykim          #+#    #+#             */
-/*   Updated: 2021/11/23 16:52:27 by taehykim         ###   ########.fr       */
+/*   Updated: 2021/12/01 11:01:24 by taehykim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (!*needle)
